Update UI, Camera, Spike and damage API
Scene and prefab updates in Kerwin.unity and related prefabs:
- UI adjustments: modified prefab UI anchors/positions/scales and added a PlayerName TextMeshPro element to the UI hierarchy.
- Camera prefab: repositioned camera, set TrackingTarget to the player, and wired cameraController references (player and cam).
- Player prefab overrides: set maxHP to 100 and adjusted Rigidbody2D gravityScale to 3.
- Spike prefab: changed Rigidbody2D body type from Dynamic (0) to Static (2).
- Scripts: changed IDamage.takeDamage parameter from int to float and updated damage implementation accordingly.
These changes fix UI layout, camera targeting, and align damage API with floating-point health values.

+++ DigitalBitSmiths - Team Project/Assets/Scripts/damage.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class damage : MonoBehaviour
 {
-    [SerializeField] int damageAmount = 10;
+    [SerializeField] float damageAmount = 10.0f;
     [SerializeField] float damageRate = 1.0f; // Seconds between hits
 
     bool isDamaging;
